+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Added support for [`3DTILES_content_conditional`](https://github.com/CesiumGS/3d-tiles/pull/834) in `Cesium3DTiles`, `Cesium3DTilesReader`, and `Cesium3DTilesWriter`.
 - Added `CesiumVectorOverlays::VectorTilesRasterOverlay` supporting vector data loaded from 3D Tiles tilesets.
 - `GeoJsonDocumentRasterOverlay` and `VectorTilesRasterOverlay` can now display points.
+- Added support for feature IDs in glTF translated from `MAXAR_content_geojson` tiles.
 
 ### v0.60.0 - 2026-05-01